Translation of the file "bird.dialog" into Russian with the help of Asd-ed.
Hello, recently in the steam under the post "Friendly Fire version 2.3.0" there was a comment about the translation of the game into other languages "especially into Russian", you threw off the site where the entire code of the game is located, where you can send your translation. I decided to translate it into Russian and tried to make it high-quality. If you are wondering why I want to translate the game into this language ,then I will answer that your game, in my opinion, is very good in the plot and in the gameplay and what I read in the reviews that there is no Russian translation. I do this translation for free. If you think that you don't really need a translation ,then I don't have to do it. And if you do not mind it, then I will gladly and diligently translate the game.